requirements.txt
fastapi
uvicorn
motor
python-dotenv
email-validator
passlib
python-jose
bcrypt==4.0.1
certifi
python-multipart